refactor: drop MCP-side org provisioning — Clerk auto-create owns it
Clerk's "create first organization automatically" instance setting now
provisions every new user's org during sign-up (enabled 2026-07-23), the
console no longer auto-creates (agentmail-web#381), and pre-existing
zero-org users were backfilled (2026-07-24). Delete the grace-poll /
single-flight / createOrganization machinery: the MCP server never
writes organizations. Zero memberships is now an anomalous state that
surfaces an actionable sign-in-at-console error.

Keep the one permanent piece: the bounded 403/404 retry in
getInternalOrganizationId that covers the seconds between org creation
and the webhook writing the AgentMail mapping — a window every fresh
signup's first tool call can legitimately hit.
